[BUGFIX] Bind onChange alert to field again
The change of #88665 aimed to centralize the onchange behavior for all
FormEngine fields. The result of this change is that this didn't work
for all field types. Due to the fact the FormEngine rendering is rather
cluttered (e.g. sometimes there are hidden fields that hold the actual
value) and it's not possible to bind a global event delegation to such
fields right now, this patch now uses the onchange handling again which
was tried to be replaced.

Resolves: #90672
Related: #88665
Releases: master
Change-Id: I17d54edfc45dd39735bcc53ed0fd92e9fa061f13
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63621
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>