[BUGFIX] Keep language or colPos when moving a record in list module
When copying a tt_content record in list module, the DataHandler would
internally look up the 'copyAfterDuplFields' settings in TCA and ensure,
that those fields are copied from its new neighboring record.

In TYPO3 core, the 'copyAfterDuplFields' settings are only used by the
table tt_content with 'colPos,sys_language_uid' as its values.

This makes sense for copying records, but the same behaviour, that stems
from TYPO3s very early days, was also used when records where moved
around with the DataHandler. When an editor changes the position of a
tt_content record in the list module, the record would then always
automatically adopt the colPos/sys_language_uid of its new neighbour,
which is usually not expected or desired, and often leads to much
confusion and broken content grids.

The code responsible for 'copyAfterDuplFields' is now removed from
DataHandler::moveRecord_raw(), while retaining the behaviour for
DataHandler::copyRecord(). Moving around tt_content records in the
page module to different columns or languages is not affected by this
change, as the colPos/sys_language_uid changes there are set in a
separate update statement.

Resolves: #72988
Resolves: #59901
Resolves: #39798
Resolves: #25216
Resolves: #14873
Releases: master, 9.5
Change-Id: Ic9c57fe2712b0996bc7b53cce4bcdc275c2820cb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64137
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>