[BUGFIX] Allow folder sorting by modification time
The LocalDriver did not allow sorting folders by modification time.
However, the property is available and can be used for sorting (for
example in the file list). This has been adjusted so sorting by
modification time is possible for both files and folders and the file
list no longer displays folders in seemingly random order.

Resolves: #76930
Releases: master, 9.5
Change-Id: Ib7711a6050ea220f015038b30607f7cac4ac8e14
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63673
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>