[BUGFIX] Provide inlineParentUid in FormInlineAjaxController
By expanding an inline record the AJAX request calls the `detailAction`
in `FormInlineAjaxController`. The parent data array created in this
function does not contain the parent record's uid nor its database row.

In order to ensure that the uid is not NULL, in any FormDataProvider,
the uid is now added to the parent data array. This value is then used
as fallback in case there is no database row given in the `compileChild`
function which calls the FormDataProviders.

Resolves: #90734
Releases: master, 9.5
Change-Id: I66a4a20980d6c62a3e87496b6261b25f5e22aea6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63976
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>