[BUGFIX] Make DateViewHelperTests resilient to execution time
Executing the tests over midnight caused the changed tests to fail
as those where comparing the "current date" from the test subject
with the one gathered within the tests, those not necessarily
are on the same day.

As the DateViewHelper uses the date from the Context we
safely can use the EXEC_TIME of the Core for the expectation.

Resolves: #90752
Releases: master, 9.5
Change-Id: I3334f0fc5c3d956a4f6801a3bcd849f7f6585497
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63700
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Susanne Moog <look@susi.dev>