[BUGFIX] Show correct value in translated single select + checkboxes
If the TCA option "l10n_mode" is set to "exclude", the selected field
in localised elements is wrong when using this option in conjunction
with "l10n_display" => "defaultAsReadonly" in a select single element.

In this case the variable $parameterArray['itemFormElValue'] seems to be
a string, not an array as expected in SelectSingleElement.php.

This patch adds a check for the array / string, as well as the
same behaviour for the checkbox.

Resolves: #76785
Resolves: #77257
Resolves: #84997
Resolves: #82843
Releases: master, 9.5
Change-Id: Ied0708dc378d73a8131791b28ebe1c61555c52aa
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63977
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>