[BUGFIX] Avoid error when using admPanel.hide = 1
The method StateUtility::isHiddenForUser must return a bool,
but instead the User TSconfig setting admPanel.hide = 1
was returned unchanged as a string. This conflicts with
the defined return type.

Additionally, a check on the backend user is added to
avoid another error when it is null.

Furthermore, missing tests in StateUtility are added.

Resolves: #90764
Related: #87231
Releases: master
Change-Id: I677fe3a808d806a1b19b53f43803d6fcaae996df
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63728
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>