[BUGFIX] Fix order of deleting page translations in DataHandler
If a page translation exists, it should be deleted before the
original page.

It the translation is deleted after the deletion of the original
page, this will result in an error because the existence and
delete permissions of the original page is checked before
deletion of the translated page.

Resolves: #85824
Releases: master, 9.5
Change-Id: I74b333fd450ea4bf361efbcd39a4daa07002c930
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63871
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>