[TASK] Render edit popup of SelectMultipleSideBySide at intuitive place
When using the edit popup of SelectMultipleSideBySide editing is related
to the items that are active and selected in the element. The old
placement of the edit pencil however suggested that it should be
possible to edit all (even inactive items, unrelated to the current
record) elements. Whenever an editPopup is configured to be shown in
a SelectMultipleSideBySide element, the edit pencil is now placed
directly next to the active select items.

Resolves: #90777
Releases: master
Change-Id: Iac8df9dafca455adb4b18d5bca882fa2ab3bd0cc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63792
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>