[BUGFIX] Never show failed "ADMCMD expired" exception
If somebody calls www.example.com?AMDCMD_prev=abcbenni the output
should never throw an exception (also if the preview entry or the
cookie has expired).

The cookie itself is a session-based cookie nowadays, so the expiration
of the cookie is not relevant.

The change lets outsiders just fall back silently to the Live version.

Resolves: #18761
Releases: master, 9.5
Change-Id: I1ca1dce2a01ce01daddc6945a45960b3e9e1f4ee
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63818
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>