[TASK] Allow workspace preview links independent of backend users
Current implementation for defining a maximum lifetime of a workspace-related
previewlink is based on userTSconfig - which actually relies on the user configuration
that creates the preview link. This can lead to massive side effects.

The option is therefore added to each workspace with a default value of "48 hours",
to properly define the lifetime on a per-workspace and not on a
"per-user-who-generated the links" basis.

Actual features will be to add a GUI to the workspace module and see currently
working preview links. The UserTSconfig option should be marked as
deprecated in further iterations together with the new GUI feature.

Resolves: #66948
Releases: master
Change-Id: I76c2dabef81d5e1b3fc2d88e049c9df09381cb6f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63881
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>