[BUGFIX] Remove mention of deprecation log in presets
The debug/live presets toggled the deprecation log in earlier versions.
Since rewriting the deprecation log to the logging framework, the
presets do not take deprecations into account anymore. Due to the
configuration merge between Default and LocalConfiguration it is
currently not possible to disable the deprecation log by overwriting
it in LocalConfiguration - setting will only take effect when set in
AdditionalConfiguration. Therefor we are removing the mention of
deprecation logs in the presets to avoid confusion.

Resolves: #89769
Releases: master, 9.5
Change-Id: I0bc1db480a31692350ea20454cfc5b143edcc70a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63895
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>