[BUGFIX] Let t3editor consume as much space as possible
T3editor now respects the configured `rows` attribute of its textarea.
If the attribute isn't configured for the node in FormEngine, the
attribute is omitted. To be able to achieve this, the enforcement of a
value for "rows" has been removed.

If the "rows" attribute is missing, the T3editor instance may consume as
much space as available, which comes in handy in the filelist.

Resolves: #86699
Releases: master, 9.5
Change-Id: Iccb152b7a747918cfc4459721c7f69a8f67cb03d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63900
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>