[TASK] Clean up drawing configuration for Fluid-based page module
PageLayoutView now builds a "Configuration" model for both variants
PageLayoutView is not set up via the DrawingConfiguration object,
which is now based on a factory method.

PageLayoutController now injects all RequireJS modules in one place,
and not in fluid anymore.

Minor bugs regarding "getLanguageColumns" are fixed.

The main "pageLayoutConfiguration" is now injected and used where it does not
differ from the backend module configuration.

Resolves: #90858
Releases: master
Change-Id: I16925488e0bff75ec73f9dddbfd047cca3bd2da1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63938
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>