[TASK] Refactor fluid page module components
Refactoring with the goal of:

* Restoring entity-like nature of BackendLayout
* Decreasing coupling between components of the page
  module processing and rendering.

Implements a PageLayoutContext to hold all context-related
objects and methods and passes the context to objects
instead of passing the BackendLayout. Reduces number of
methods on DrawingConfiguration so the configuration object
only holds output- and behavior-related options and does not
hold things like permission checks.

Releases: master
Resolves: #90882
Change-Id: I09f1d97a53a76ea96d6cbc18bb5549127b5a1a71
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64002
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Benni Mack <benni@typo3.org>