[TASK] Remove dependency to algo26-matthias/idna-convert
The third-party library algo26-matthias/idna-convert is not compatible
with PHP 7.4 (used for providing IDNA conversions for umlaut domains).

The classes from v1 are copied in order to have them still available, but we
maintain the code ourselves and reduce the usage to a minimum.

  composer remove "algo26-matthias/idna-convert:"

See https://idnaconv.net/docs.html for detailed upgrade
changes.

Resolves: #90911
Releases: 9.5
Change-Id: I7ac14348677b8b60f92cb8bfa26e26d16a01abee
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64011
Tested-by: Manuel Selbach <manuel_selbach@yahoo.de>
Tested-by: Frank Nägler <frank.naegler@typo3.org>
Reviewed-by: Manuel Selbach <manuel_selbach@yahoo.de>
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>