[TASK] Migrate checkPidArray into PageRepository
The code in cObj->checkPidArray() and cObj->checkPid() is checking
for valid page records, which semantically belongs to PageRepository.

As this code is now added to PageRepository, it is built in a more
flexible way now, and available for all core relevant parts, not
just Frontend logic.

The method cObj->checkPid(), previously marked as internal, is removed.

Resolves: #90928
Releases: master
Change-Id: I4e6c089ed7efcab208a1aa3bb79599cf2c719787
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64042
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>