[BUGFIX] correctly resolve saved mm relations in TCA
Using mm tables, the value that is stored in the database does not represent the
uid of the selected item. Much rather it is the amount of selected items.
If three items are selected the value within the database field is "3".
Whereas using multiple select without mm table, it is stored as a
comma-separated list of uids.

Releases: master, 9.5
Resolves: #90925
Change-Id: I110cbf0fd19fab1135958c445a846a1af9c24510
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64134
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>