[TASK] Deprecate runtime activated packages
The option to register packages during runtime was a work-around
to dynamically modify the "extList". Using this feature has certain limitations:

* Runtime-activated Extensions cannot add their DI configuration
* Runtime-activated Extensions make every (!) single TYPO3 request much slower
  just like back in 6.2.0 times

The usecases from people using this functionality was to enable e.g. extensions
such as "devlog", "mask"/"mask_export" or "extensionbuilder" only on development
systems.

Nowadays, it is recommended - if this functionality is needed - to use TYPO3
Console and Composer Mode (with require-dev) to achieve a similar behaviour.

This also ensures a clear state of available extensions, and to build a proper
class map / autoloading mechansims of extensions in use.

Runtime-activated extensions also prohibit to continue to work on
a smoother integration between composer-mode and non-composer mode
in the future.

Resolves: #91030
Releases: master
Change-Id: If4344cd081e8f60f30c9a0fb4d0cb9c308bbfb66
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64166
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Helmut Hummel <typo3@helhum.io>
Tested-by: Christian Eßl <indy.essl@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Josef Glatz <josefglatz@gmail.com>
Reviewed-by: Christian Eßl <indy.essl@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>