[TASK] Restructure and fine-tune builds
This patch updates the Docker configuration to use mssql 2019 as mssql
2017 images don't get updates anymore.

Additionally to the mssql update, the shuffled chunking to 50 jobs is
reverted as it didn't bring any benefit but even more hassle as the
stages failed earlier and more often. Instead, the jobs are now grouped
by PHP version which allows smaller chunks and reduced the probability to
clog the build queue.

To workaround errors thrown by Gerrit on heavy load, the cherry-pick
command is now executed up to five times, with a short sleep between each
execution.

Last but not least we check whether there is actually a dangling Docker
container instead of hard killing non-existent containers which pollutes
our log.

Resolves: #90965
Releases: master, 9.5
Change-Id: Ied60297127b86c3375d310759bf441ad9e10aeed
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64087
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>