[BUGFIX] Load the correct "first module"
The start up option "First module in menu" is a misleading in some ways.
The select box e.g. has "dashboard" as first selectable option, but in
such cases the "help_AboutAbout" module is loaded instead, for example.
This is due the fact the first module is determined via querying the DOM
for the first element matching the selector
`.t3js-modulemenu-action[data-link]`.

The behavior is now changed to load the first module as listed in the
previously mentioned select box.

Resolves: #91028
Releases: master
Change-Id: I669a9b146f76f2513c51ff5f48128a1ea8b0c0a7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64164
Tested-by: Richard Haeser <richard@maxserv.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Richard Haeser <richard@maxserv.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>