[BUGFIX] Provide symfony container in install tool "Check TCA" action
A booted symfony container is needed as ext_localconf.php is
loaded since #89713 for the "Check TCA" action.
ext_localconf.php files may depend on a fully-booted
symfony container, therefore we need to provide it in this
case as well – or ext_localconf.php loading may fail.

Adapt the LateBootService to allow to reset to a
previously backuped container (even if it has been
reset in the meantime by another method).

Currently affected extensions were EXT:news before commit
ac0cb20c (which added a workaround for this case).
Thus the bugfix can be verified against the parent of that commit:
georgringer/news:205cb3e

Releases: master
Resolves: #91073
Related: #89713
Change-Id: Iaceb5efae26c1ae2e08e0db31ff2c223cc803ce7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64190
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Josef Glatz <josefglatz@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>