[TASK] Mark various functionality in TSFE internal
Some code related to templating / cached TypoScript,
inclusion of CSS/JS and internals like "indexedDocTitle"
which has now a proper PageTitle API, is marked as
internal now, so further refactorings can happen
in the next major release.

As all replacements are in place for this functionality,
extension authors can and should already use new APIs
instead of relying on TSFE for this logic.

Resolves: #91079
Releases: master
Change-Id: Ic24c607dd620d1cfe7966484d5272491aa5f429a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64199
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>