[BUGFIX] Boot container for E-Mail check in install tool
Since #90266 install tool E-Mails are templated and
custom logos, viewhelpers or services may be loaded.

A custom backend login logo caused the usage of the f:image
ViewHelper which triggered the FileIndexRepository which
wasn't configured for the install tool context.

Therefore we need to provide a fully booted
symfony container in order for dependency injection
to work for all core and third party extensions,
especially when using a customized SystemEmail template.

Resolves: #91067
Releases: master
Change-Id: Ia2cb06dfecd7bb9bd757b37849c1345e538dfffc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64205
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Martin Hotmann <martinhotmann@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Martin Hotmann <martinhotmann@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>