[FEATURE] Add 'show scheduled records' option to admin panel
With this new option an editor is able to view all records on
a page regardless of the start and end time.

This is especially helpful if one has content on a page with different
start and end times and an editor wants to have an overview of all
elements on that page. For example if he wants to check if the layout
and everything is okay. Without this option the editor has to simulate
multiple dates to be able to see all records, because records which are
visible on one date, can be invisible on another.

Resolves: #91082
Releases: main
Change-Id: If4bd4103615f17a17fac8cb003e439bee65301da
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64208
Tested-by: core-ci <typo3@b13.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Benni Mack <benni@typo3.org>