[BUGIX] Register MetaDataRepository via service provider
The installtool sends a mail when a login happens when
TYPO3_CONF_VARS.BE.warning_email_addr is set.
When TYPO3_CONF_VARS.EXTENSIONS.backend.loginLogo is set
to provide a custom backend logo, that logo is used by
the FluidEmail template via f:image viewhelper.

Therefore we need to provide MetaDataRepository,
as the f:image viewhelper will use that service.
The (implicit) dependency chain is:

  <f:image>
    -> Resource\File
      -> MetaDataAspect
        -> MetaDataRepository

Releases: master
Resolves: #91107
Related: #89892
Change-Id: If4e73d607d2a6e21646ea91bd8f0a39742f133d4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64236
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Jonas Eberle <flightvision@googlemail.com>
Tested-by: Alexander Schnitzler <git@alexanderschnitzler.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Alexander Schnitzler <git@alexanderschnitzler.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>