[BUGFIX] Allow more characters for MySQL / MariaDB database name
With this change it is possible to use a wider set of characters.
Please have a look at the official documentation of MySQL / MariaDB.

e.g.:

https://dev.mysql.com/doc/refman/5.5/en/identifiers.html
https://dev.mysql.com/doc/refman/5.6/en/identifiers.html
https://dev.mysql.com/doc/refman/5.7/en/identifiers.html
https://dev.mysql.com/doc/refman/8.0/en/identifiers.html
https://mariadb.com/kb/en/identifier-names/

The mentioned characters in chapter "Quoted" (ASCII and Extended) are
supported by now.

Furthermore the database name is quoted during the installation process
for creating / dropping a database.

If, for example a `.` is used in the name, a notification will be shown.

e.g.:

```
Unable to create database
Database with name "foo.test@bla123" could not be created. Either your database name contains a reserved keyword or your database user does not have sufficient permissions to create it or the database already exists. Please choose an existing (empty) database, choose another name or contact administration.
```

Resolves: #91167
Releases: master
Change-Id: I2df93f5c2238c2f0ca5ab8020ca8eebd10fdf58f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64312
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Manuel Selbach <manuel_selbach@yahoo.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Manuel Selbach <manuel_selbach@yahoo.de>